unify db test, test no auth to api
